back/pyproject.toml
[tool.poetry]
name = "back"
version = "0.1.0"
description = ""
authors = ["т"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
anyio = "^4.6.2.post1"
attrs = "^24.2.0"
build = "^1.2.2.post1"
cachecontrol = "^0.14.1"
camelot-py = "^0.11.0"
certifi = "^2024.8.30"
cffi = "^1.17.1"
chardet = "^5.2.0"
charset-normalizer = "^3.4.0"
cleo = "^2.1.0"
click = "^8.1.7"
crashtest = "^0.4.1"
cryptography = "^43.0.3"
distlib = "^0.3.9"
distro = "^1.9.0"
dnspython = "^2.7.0"
dulwich = "^0.22.6"
email-validator = "^2.2.0"
et-xmlfile = "^2.0.0"
fastapi = "^0.115.5"
filelock = "^3.16.1"
html5lib = "^1.1"
httptools = "^0.6.4"
idna = "^3.10"
installer = "^0.7.0"
jaraco-classes = "^3.4.0"
jdcal = "^1.4.1"
jeepney = "^0.8.0"
jsonschema = "^4.23.0"
keyring = "^25.5.0"
lxml = "^5.3.0"
more-itertools = "^10.5.0"
motor = "^3.6.0"
msgpack = "^1.1.0"
numpy = "^2.1.3"
opencv-python = "^4.10.0.84"
openpyexcel = "^2.5.14"
openpyxl = "^3.1.5"
pandas = "^2.2.3"
pdfminer-six = "^20240706"
pexpect = "^4.9.0"
pypdf2 = "^3.0.1"
python-docx = "^1.1.2"
python-multipart = "^0.0.17"
pytz = "^2024.2"
pyyaml = "^6.0.2"
rapidfuzz = "^3.10.1"
requests = "^2.32.3"
requests-toolbelt = "^1.0.0"
secretstorage = "^3.3.3"
six = "^1.16.0"
sniffio = "^1.3.1"
tabula-py = "^2.10.0"
tzdata = "^2024.2"
urllib3 = "^2.2.3"
uvicorn = "^0.32.0"
uvloop = "^0.21.0"
watchfiles = "^0.24.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.36"}
asyncpg = "^0.30.0"
pydantic-settings = "^2.7.0"
redis = "^5.2.1"
pyjwt = "^2.10.1"
alembic = "^1.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
